Add checks for the data files and test all checks (#2)
I chose to leave the folio check using the built-in python http client. It isn't great but we can put off committing to choosing a library for a while longer. While adding actual tests for it I fixed some bugs and edge cases and refactored it to work better.

For the data we're checking both that we can read it and that it matches a schema. The schema checks are super basic right now and I'll focus on adding more checks once we have more requirements. I chose to use Polars because I've worked with it before and like how it handles datatypes using pyarrow instead of numpy and has a better api than pandas. I chose to use a library rather than the build in CSV reader because I don't want to deal with all the conversion and reading and data cleanup issues myself. Pandera plays nicely with polars and let's us specify pretty broad checks on the data.