Add cli and plumbing (#3)
Instead of taking a dependency on a CLI library from the beginning I chose to just use argparse. In the future when we add more subcommands we can either switch to a library or add subcommands and groups to keep it organized.

The logging is more or less hardcoded to log basic information to a file and the console for ease of human use, with more information being logged as a tsv to a file if we really need to troubleshoot. I didn't add any log statements right now. They can be added in the future as we work on these modules more.

Additionally I setup the convenience bits for using this as a cli or executing it as a module.