Add command to cli for importing users (#6)
I used argparse subparsers for the two commands. There is some goofiness with how subparsers interact with nargs and how the help is displayed. I got around it with reading the files in two parameters which isn't great but it's documented.

I copied the pattern used for the check command. The cli file is getting pretty messy but I don't think it will keep getting messier. Maybe it should be refactored.