Refactor commands to be able to share code (#7)
Initially I kept it simple by putting all the code for check in one place. Now the user import command will need access to a lot of the same plumbing. I built out a poc import command to get a sense of what was necessary and then moved all the code into the root of the package.

I'm not super sure if all the guts should be exposed like this but it does seem somewhat useful to packages consumers. I also am not sure if the test code should go in the base module or in the check module but it easiest this way.

Someone in FOLIO slack is working on https://github.com/balljok/pyfolioclient . It seems to do what we need and would save re-implementing this using httpx directly. I chose to switch to use this as the dependency for now, we can always go back to a more direct implementation if necessary.