Import users into Folio (#8)
This stitches everything together so far by transforming the flat csvs into request bodies and sending them off to the user-import endpoint in batches. I'm not super happy with the organization of the transformations but it is well black-box tested for now and can be refactored later if neccessary. I ended up reusing the samples from the check command to test the transformations, there's maybe a better way to organize those sample files. I removed two options for import that were aspirational at best and no one really asked for them. The performance seems fine so far but we can add concurrency back in the future.

Notably, this doesn't collect any of the reasons a user failed to import. This PR was already getting pretty sizeable so I'm going to add it as part of another one.