Rename from fuiman to ube (#10)
This tool is going to have bulk anonymization and deletions added to it. I'm changing the name to be less user import specific. I've been made aware that the existing bulk edit in FOLIO technically supports users but I don't think there will be confusion as the existing bulk edit doesn't really do anything. I may just be too attached to the name ube for the cli though.