Check for users that are active but have a past expiration date (#11)
While testing the import I noticed that there are users which are both active and have an expiration date in the past. They get imported as active but then FOLIO deactivates them shortly after. This isn't a validation issue but a more of a warning? I wonder if this should be raised separately somehow.